requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
docutils==0.20.1
gpxpy==1.5.0
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
statistics==1.0.3.5
urllib3==2.0.6